+++ src/test/fuzz/util.h
@@ -251,7 +251,7 @@ CNetAddr ConsumeNetAddr(FuzzedDataProvider &fuzzed_data_provider) noexcept {
 
 CSubNet ConsumeSubNet(FuzzedDataProvider &fuzzed_data_provider) noexcept {
     return {ConsumeNetAddr(fuzzed_data_provider),
-            fuzzed_data_provider.ConsumeIntegral<int32_t>()};
+            fuzzed_data_provider.ConsumeIntegral<uint8_t>()};
 }
 
 void InitializeFuzzingContext(
